requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
playwright==1.44.0
httpx==0.27.0
python-whois==0.9.4
dnspython==2.6.1
Pillow==10.3.0
imagehash==4.3.1
jieba==0.42.1
transformers==4.41.0
torch==2.3.0
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1
aiofiles==23.2.1
cryptography==42.0.7
beautifulsoup4==4.12.3
lxml==5.2.2
tqdm==4.66.4
loguru==0.7.2
redis==5.0.4
celery==5.4.0
